refactor: Update useUserDetails hook to include additional user details
This commit modifies the useUserDetails hook in the src/shared/hooks/useUserDetails.jsx file to include additional user details such as name, lastName, email, rolerUser, numbercel, and address. The hook now returns these details along with the existing isLogged and username properties. This change improves the functionality of the hook and allows for more comprehensive user information to be accessed.